requirements.txt
anyio==3.7.0
certifi==2023.5.7
click==8.1.3
exceptiongroup==1.1.1
fastapi==0.95.2
h11==0.14.0
httpcore==0.17.2
httptools==0.5.0
httpx==0.24.1
idna==3.4
iniconfig==2.0.0
loguru==0.7.0
packaging==23.1
pluggy==1.0.0
psycopg2-binary==2.9.2
pydantic==1.10.8
pyte==0.8.1
pytest==7.3.1
python-dotenv==1.0.0
PyYAML==6.0
sniffio==1.3.0
starlette==0.27.0
tomli==2.0.1
typing_extensions==4.6.2
uvicorn==0.22.0
uvloop==0.17.0
watchfiles==0.19.0
wcwidth==0.2.6
websockets==11.0.3
